Create deprecation warnings for old flag functions
And fix an inaccuracy in one of the doc comments
