+++ packages/design-system/src/components/config.ts
@@ -39,8 +39,7 @@ export interface Config {
    */
   helpDrawerSendsAnalytics: boolean;
   /**
-   * Controls whether the header component send analytics data by default. Alternatively,
-   * this can be set on the header instance itself using the `analytics` prop.
+   * Controls whether the header component send analytics data.
    */
   headerSendsAnalytics: boolean;
 }
@@ -74,13 +73,67 @@ config.HEALTHCARE_DEFAULTS = HEALTHCARE_DEFAULTS;
 
 // Deprecated stuff
 
-export function setDefaultAnalyticsFunction(analyticsFunction: AnalyticsFunction) {
+function depWarning(fnName: string) {
   if (process.env.NODE_ENV !== 'production') {
     console.warn(
-      'The `setDefaultAnalyticsFunction` function is deprecated. Please use `config` instead.'
+      `The '${fnName}' function is deprecated. Please use the global 'config' function instead.`
     );
   }
+}
+
+export function setDefaultAnalyticsFunction(analyticsFunction: AnalyticsFunction) {
+  depWarning('setDefaultAnalyticsFunction');
   config({ defaultAnalyticsFunction: analyticsFunction });
 }
 
 export const defaultAnalyticsFunction = config().defaultAnalyticsFunction;
+
+export function errorPlacementDefault(): ErrorPlacement {
+  depWarning('setDefaultAnalyticsFunction');
+  return config().errorPlacementDefault;
+}
+
+export function setErrorPlacementDefault(value: ErrorPlacement): void {
+  depWarning('setErrorPlacementDefault');
+  config({ errorPlacementDefault: value });
+}
+
+export function alertSendsAnalytics(): boolean {
+  depWarning('alertSendsAnalytics');
+  return config().alertSendsAnalytics;
+}
+
+export function setAlertSendsAnalytics(value: boolean): void {
+  depWarning('setAlertSendsAnalytics');
+  config({ alertSendsAnalytics: value });
+}
+
+export function buttonSendsAnalytics(): boolean {
+  depWarning('buttonSendsAnalytics');
+  return config().buttonSendsAnalytics;
+}
+
+export function setButtonSendsAnalytics(value: boolean): void {
+  depWarning('setButtonSendsAnalytics');
+  config({ buttonSendsAnalytics: value });
+}
+
+export function dialogSendsAnalytics(): boolean {
+  depWarning('dialogSendsAnalytics');
+  return config().dialogSendsAnalytics;
+}
+
+export function setDialogSendsAnalytics(value: boolean): void {
+  depWarning('setDialogSendsAnalytics');
+  config({ dialogSendsAnalytics: value });
+}
+
+export function helpDrawerSendsAnalytics(): boolean {
+  depWarning('helpDrawerSendsAnalytics');
+  return config().helpDrawerSendsAnalytics;
+}
+
+export function setHelpDrawerSendsAnalytics(value: boolean): void {
+  depWarning('setHelpDrawerSendsAnalytics');
+  config({ helpDrawerSendsAnalytics: value });
+}
