+++ packages/design-system/src/components/index.ts
@@ -32,6 +32,5 @@ export * from './Icons';
 
 export * from './analytics';
 export * from './config';
-export * from './flags';
 export * from './i18n';
 export { default as useId } from './utilities/useId';
